requirements.txt
streamlit
numpy
opencv-python-headless
pillow
scikit-learn
collections